Change rating type to integer and add example for review_form data

+++ models/review.go
@@ -1,11 +1,11 @@
 package models
 
 type Review struct {
-	Username    string  `json:"username"`
-	PicturePath string  `json:"picture_path"`
-	Rating      float32 `json:"Rating"`
-	Content     string  `json:"content"`
-	Date        string  `json:"date"`
+	Username    string `json:"username"`
+	PicturePath string `json:"picture_path"`
+	Rating      int    `json:"Rating"`
+	Content     string `json:"content"`
+	Date        string `json:"date"`
 }
 
 type Reviews struct {
